broke MRSortAndGroupReversedText into its own file
